properly parse 'optional' uniforms into material properties add support for multiple projection maps in both Built-in shaders fix building config file class on mobile

+++ common/src/ecs/systems/TextRenderer.cpp
@@ -359,7 +359,7 @@ void TextRenderer::rebuildBatch()
             }
 
             //increase the start point
-            start += text.m_vertices.size();
+            start += static_cast<uint32>(text.m_vertices.size());
 
             //and append data
             text.m_vboOffset = static_cast<int32>(vertexData.size() * sizeof(float));
